Add xfail test for heterogenous dicts in a list.
dget shouldn't assume all dicts in a list have the same structure. However, this may have
wider implications. A design decision is required as this behaviour is now becoming controversial
since if you dget a list with 5 dicts but the key exists in 3 out of 5, how long should the list
be. If you say three well how do you know which value belongs to which dict, if you say five then
we need to design on a default value.

I believe #8 may cause a regression in the current behaviour.